Solved InsufficientPermissionException(Permission Missing Error).

+++ src/main/java/com/DecupleProject/Listener/DefaultListener.java
@@ -236,7 +236,7 @@ public class DefaultListener extends ListenerAdapter {
 
                 if (se.hasEmoji(name) && msg.getContentRaw().contains(code)) {
                     se.sendStealEmoji(name);
-                    tc.deleteMessageById(msg.getId()).queue();
+                    if (Objects.requireNonNull(guild.getMember(DefaultListener.jda.getSelfUser())).hasPermission(Permission.MESSAGE_MANAGE)) tc.deleteMessageById(msg.getId()).queue();
                 }
             }
 
@@ -249,7 +249,7 @@ public class DefaultListener extends ListenerAdapter {
                 if (e.eq(args[0], "도움말") | e.eq(args[0], "도움") | e.eq(args[0], "help")) {
 
                     if (user.isBot()) return;
-                    tc.deleteMessageById(msg.getId()).queue();
+                    if (Objects.requireNonNull(guild.getMember(DefaultListener.jda.getSelfUser())).hasPermission(Permission.MESSAGE_MANAGE)) tc.deleteMessageById(msg.getId()).queue();
                     eb.setTitle("도움말을 보여드릴게요!");
                     eb.setDescription("아쉽지만, 명령어가 너무 많아서 정리를 해놓은 인터넷 사이트로 이동시켜 드리겠습니다!");
                     eb.addField("링크", "https://decupleproject.fandom.com/ko/wiki/%EB%8D%B0%ED%81%90%ED%94%8C%ED%94%84%EB%A1%9C%EC%A0%9D%ED%8A%B8:%EB%94%94%EC%8A%A4%EC%BD%94%EB%93%9C%ED%80%B8%ED%8A%9C%ED%94%8C", true);
@@ -262,7 +262,7 @@ public class DefaultListener extends ListenerAdapter {
                 if (e.eq(args[0], "핑") | e.eq(args[0], "ping")) {
 
                     if (user.isBot()) return;
-                    tc.deleteMessageById(msg.getId()).queue();
+                    if (Objects.requireNonNull(guild.getMember(DefaultListener.jda.getSelfUser())).hasPermission(Permission.MESSAGE_MANAGE)) tc.deleteMessageById(msg.getId()).queue();
                     long time = System.currentTimeMillis();
 
                     tc.sendMessage("불러오는 중입니다!").queue(response -> {
@@ -283,14 +283,14 @@ public class DefaultListener extends ListenerAdapter {
                 if (e.eq(args[0], "초대") | e.eq(args[0], "invite")) {
 
                     if (user.isBot()) return;
-                    tc.deleteMessageById(msg.getId()).queue();
+                    if (Objects.requireNonNull(guild.getMember(DefaultListener.jda.getSelfUser())).hasPermission(Permission.MESSAGE_MANAGE)) tc.deleteMessageById(msg.getId()).queue();
                     sendPrivateMessage(user, "이 링크를 이용하시면, 데큐플 공식 서버로 가입하실 수 있어요! <http://decuple-d.o-r.kr/>");
 
                 }
 
                 if (e.eq(args[0], "시간") | e.eq(args[0], "시각") | e.eq(args[0], "time") | e.eq(args[0], "t")) {
 
-                    tc.deleteMessageById(msg.getId()).queue();
+                    if (Objects.requireNonNull(guild.getMember(DefaultListener.jda.getSelfUser())).hasPermission(Permission.MESSAGE_MANAGE)) tc.deleteMessageById(msg.getId()).queue();
                     SimpleDateFormat format1 = new SimpleDateFormat("yyyy년 MM월 dd일");
                     SimpleDateFormat format2 = new SimpleDateFormat("HH시 mm분 ss초");
 
@@ -316,7 +316,7 @@ public class DefaultListener extends ListenerAdapter {
 
                 if (e.eq(args[0], "봇", "Bot", "정보")) {
 
-                    tc.deleteMessageById(msg.getId()).queue();
+                    if (Objects.requireNonNull(guild.getMember(DefaultListener.jda.getSelfUser())).hasPermission(Permission.MESSAGE_MANAGE)) tc.deleteMessageById(msg.getId()).queue();
                     ReadFile r = new ReadFile();
 
                     eb.setTitle("봇 : 퀸튜플");
@@ -370,7 +370,7 @@ public class DefaultListener extends ListenerAdapter {
                 }
 
                 if (e.eq(args[0], "계산") | e.eq(args[0], "Calculate") | e.eq(args[0], "Calculating") | e.eq(args[0], "Calculator") | e.eq(args[0], "calc") | e.eq(args[0], "계산기")) {
-                    tc.deleteMessageById(msg.getId()).queue();
+                    if (Objects.requireNonNull(guild.getMember(DefaultListener.jda.getSelfUser())).hasPermission(Permission.MESSAGE_MANAGE)) tc.deleteMessageById(msg.getId()).queue();
                     if (user.isBot()) return;
 
                     if (args.length < 2) {
@@ -486,7 +486,7 @@ public class DefaultListener extends ListenerAdapter {
                 }
 
                 if (e.eq(args[0], "번역", "translate", "번역기", "파파고", "통역", "통역기")) {
-                    tc.deleteMessageById(msg.getId()).queue();
+                    if (Objects.requireNonNull(guild.getMember(DefaultListener.jda.getSelfUser())).hasPermission(Permission.MESSAGE_MANAGE)) tc.deleteMessageById(msg.getId()).queue();
                     try {
                         String body = String.join(" ", Arrays.copyOfRange(args, 3, args.length));
                         new Translator(body, tc, args[1], args[2]);
@@ -500,13 +500,13 @@ public class DefaultListener extends ListenerAdapter {
                 }
 
                 if (e.eq(args[0], "출석", "출석체크", "출첵", "Attendance", "AttendanceCheck", "나님등장")) {
-                    tc.deleteMessageById(msg.getId()).queue();
+                    if (Objects.requireNonNull(guild.getMember(DefaultListener.jda.getSelfUser())).hasPermission(Permission.MESSAGE_MANAGE)) tc.deleteMessageById(msg.getId()).queue();
                     AttendanceCheck at = new AttendanceCheck(user.getId(), tc, user.getName());
                     at.attendance();
                 }
 
                 if (e.eq(args[0], "백과사전", "네이버백과", "백과")) {
-                    tc.deleteMessageById(msg.getId()).queue();
+                    if (Objects.requireNonNull(guild.getMember(DefaultListener.jda.getSelfUser())).hasPermission(Permission.MESSAGE_MANAGE)) tc.deleteMessageById(msg.getId()).queue();
                     new Encyclopedia(String.join(" ", Arrays.copyOfRange(args, 1, args.length)), tc);
                 }
 
@@ -514,7 +514,7 @@ public class DefaultListener extends ListenerAdapter {
 
                     if (args.length != 1) return;
 
-                    tc.deleteMessageById(msg.getId()).queue();
+                    if (Objects.requireNonNull(guild.getMember(DefaultListener.jda.getSelfUser())).hasPermission(Permission.MESSAGE_MANAGE)) tc.deleteMessageById(msg.getId()).queue();
                     eb.setTitle("유저 : " + user.getAsTag());
 
                     eb.addField("닉네임", user.getName(), true);
@@ -549,7 +549,7 @@ public class DefaultListener extends ListenerAdapter {
                 }
 
                 if (e.eq(args[0], "보고", "report", "리폿", "버그", "이슈", "issue", "bug")) {
-                    tc.deleteMessageById(msg.getId()).queue();
+                    if (Objects.requireNonNull(guild.getMember(DefaultListener.jda.getSelfUser())).hasPermission(Permission.MESSAGE_MANAGE)) tc.deleteMessageById(msg.getId()).queue();
                     try {
                         eb.setTitle("누군가로부터 도착한 보고서입니다!");
                         eb.addField("보고하신 분", user.getAsTag(), false);
@@ -568,7 +568,7 @@ public class DefaultListener extends ListenerAdapter {
                 }
 
                 if (e.eq(args[0], "멜론차트", "인기차트")) {
-                    tc.deleteMessageById(msg.getId()).queue();
+                    if (Objects.requireNonNull(guild.getMember(DefaultListener.jda.getSelfUser())).hasPermission(Permission.MESSAGE_MANAGE)) tc.deleteMessageById(msg.getId()).queue();
                     try {
                         Melon melon = new Melon();
 
@@ -591,13 +591,13 @@ public class DefaultListener extends ListenerAdapter {
                 }
 
                 if (e.eq(args[0], "쇼핑", "shopping", "최저가", "구매")) {
-                    tc.deleteMessageById(msg.getId()).queue();
+                    if (Objects.requireNonNull(guild.getMember(DefaultListener.jda.getSelfUser())).hasPermission(Permission.MESSAGE_MANAGE)) tc.deleteMessageById(msg.getId()).queue();
                     Shopping sh = new Shopping(user.getId(), String.join(" ", Arrays.copyOfRange(args, 1, args.length)));
                     sh.sendShopMessage(tc);
                 }
 
                 if (e.eq(args[0], "DB", "데이터베이스", "database")) {
-                    tc.deleteMessageById(msg.getId()).queue();
+                    if (Objects.requireNonNull(guild.getMember(DefaultListener.jda.getSelfUser())).hasPermission(Permission.MESSAGE_MANAGE)) tc.deleteMessageById(msg.getId()).queue();
                     if (e.eq(args[1], "del")) {
                         db.editDatabase(args[1], String.join(" ", Arrays.copyOfRange(args, 2, args.length)), true);
                     }
@@ -606,7 +606,7 @@ public class DefaultListener extends ListenerAdapter {
                 }
 
                 if (e.eq(args[0], "사전", "국어사전", "국어", "뜻")) {
-                    tc.deleteMessageById(msg.getId()).queue();
+                    if (Objects.requireNonNull(guild.getMember(DefaultListener.jda.getSelfUser())).hasPermission(Permission.MESSAGE_MANAGE)) tc.deleteMessageById(msg.getId()).queue();
 
                     if (args.length == 1) {
                         eb.setTitle("뜻을 찾으려면..");
@@ -640,7 +640,7 @@ public class DefaultListener extends ListenerAdapter {
                 }
 
                 if (e.eq(args[0], "auth", "권한", "authorization")) {
-                    tc.deleteMessageById(msg.getId()).queue();
+                    if (Objects.requireNonNull(guild.getMember(DefaultListener.jda.getSelfUser())).hasPermission(Permission.MESSAGE_MANAGE)) tc.deleteMessageById(msg.getId()).queue();
                     try {
 
                         Authority a = new Authority();
@@ -680,7 +680,7 @@ public class DefaultListener extends ListenerAdapter {
                 }
 
                 if (e.eq(args[0], "한영변환", "gksdud", "en", "e2k")) {
-                    tc.deleteMessageById(msg.getId()).queue();
+                    if (Objects.requireNonNull(guild.getMember(DefaultListener.jda.getSelfUser())).hasPermission(Permission.MESSAGE_MANAGE)) tc.deleteMessageById(msg.getId()).queue();
                     if (args.length == 1) {
                         eb.setTitle("무엇을 보낼 건가요?");
                         eb.setDescription("`.en dkssudgktpdy?`와 같은 형식으로 입력해 주세요.");
@@ -753,7 +753,7 @@ public class DefaultListener extends ListenerAdapter {
                 }
 
                 if (e.eq(args[0], "Emote", "Emoji", "Emoticon", "EM", "이모티콘")) {
-                    tc.deleteMessageById(msg.getId()).queue();
+                    if (Objects.requireNonNull(guild.getMember(DefaultListener.jda.getSelfUser())).hasPermission(Permission.MESSAGE_MANAGE)) tc.deleteMessageById(msg.getId()).queue();
                     try {
 
                         StealEmoji se = new StealEmoji(user.getId(), event.getJDA(), tc);
@@ -802,7 +802,7 @@ public class DefaultListener extends ListenerAdapter {
 
                 if (e.eq(args[0], "배송", "택배", "Delivery", "운송장", "배송", "운송장번호", "운송장조회", "배송조회")) {
 
-                    tc.deleteMessageById(msg.getId()).queue();
+                    if (Objects.requireNonNull(guild.getMember(DefaultListener.jda.getSelfUser())).hasPermission(Permission.MESSAGE_MANAGE)) tc.deleteMessageById(msg.getId()).queue();
                     Delivery dv = new Delivery(tc, args[1], args[2]);
 
                     if (args.length != 3) {
@@ -822,7 +822,7 @@ public class DefaultListener extends ListenerAdapter {
                 }
 
                 if (e.eq(args[0], "날씨", "weather", "웨더")) {
-                    tc.deleteMessageById(msg.getId()).queue();
+                    if (Objects.requireNonNull(guild.getMember(DefaultListener.jda.getSelfUser())).hasPermission(Permission.MESSAGE_MANAGE)) tc.deleteMessageById(msg.getId()).queue();
                     Weather w = new Weather();
 
                     w.sendWeatherInformation(args[1], tc);
