+++ src/main/java/com/DecupleProject/Listener/OnlineGameListener.java
@@ -59,7 +59,7 @@ public class OnlineGameListener extends ListenerAdapter {
 
                 if (e.eq(args[0], "LL", "LOL", "롤", "LeagueOfLegends", "리그오브레전드")) {
 
-                    tc.deleteMessageById(msg.getId()).queue();
+                    if (Objects.requireNonNull(guild.getMember(DefaultListener.jda.getSelfUser())).hasPermission(Permission.MESSAGE_MANAGE)) tc.deleteMessageById(msg.getId()).queue();
 
                     if (e.eq(args[1], "np")) {
                         String inputted = String.join(" ", Arrays.copyOfRange(args, 2, args.length));
@@ -77,7 +77,7 @@ public class OnlineGameListener extends ListenerAdapter {
 
                 if (e.eq(args[0], "Osu", "오스", "Osu!", "오스!")) {
 
-                    tc.deleteMessageById(msg.getId()).queue();
+                    if (Objects.requireNonNull(guild.getMember(DefaultListener.jda.getSelfUser())).hasPermission(Permission.MESSAGE_MANAGE)) tc.deleteMessageById(msg.getId()).queue();
 
                     String n = String.join(" ", Arrays.copyOfRange(args, 1, args.length));
                     new Osu(n, tc);
@@ -86,7 +86,7 @@ public class OnlineGameListener extends ListenerAdapter {
 
                 if (e.eq(args[0], "오버워치", "Overwatch")) {
 
-                    tc.deleteMessageById(msg.getId()).queue();
+                    if (Objects.requireNonNull(guild.getMember(DefaultListener.jda.getSelfUser())).hasPermission(Permission.MESSAGE_MANAGE)) tc.deleteMessageById(msg.getId()).queue();
 
                     String tag = String.join(" ", Arrays.copyOfRange(args, 1, args.length));
                     Overwatch overwatch = new Overwatch(tag);
