+++ src/main/java/com/DecupleProject/Listener/ServerManagementListener.java
@@ -84,7 +84,7 @@ public class ServerManagementListener extends ListenerAdapter {
                 ServerManager manager = new ServerManager(guild, member);
 
                 if (e.eq(args[0], "청소", "cl", "cls", "clean", "삭제")) {
-                    tc.deleteMessageById(msg.getId()).queue();
+                    if (Objects.requireNonNull(guild.getMember(DefaultListener.jda.getSelfUser())).hasPermission(Permission.MESSAGE_MANAGE)) tc.deleteMessageById(msg.getId()).queue();
                     if (args.length == 1) {
                         eb.setDescription("몇 개의 메시지를 삭제하고 싶으신지 말해주세요.");
                         tc.sendMessage(eb.build()).delay(10, TimeUnit.SECONDS).flatMap(Message::delete).queue();
@@ -94,7 +94,7 @@ public class ServerManagementListener extends ListenerAdapter {
                 }
 
                 if (e.eq(args[0], "서버", "server")) {
-                    tc.deleteMessageById(msg.getId()).queue();
+                    if (Objects.requireNonNull(guild.getMember(DefaultListener.jda.getSelfUser())).hasPermission(Permission.MESSAGE_MANAGE)) tc.deleteMessageById(msg.getId()).queue();
 
                     if (args.length == 1) {
                         eb.setTitle("서버 : " + guild.getName());
